Cleanup: Move remaining NPF files into pathfinder directory
