+++ src/pathfinder/aystar.cpp
