--- src/pathfinder/npf/CMakeLists.txt
@@ -1,6 +1,0 @@
-add_files(
-    aystar.cpp
-    aystar.h
-    queue.cpp
-    queue.h
-)
