+++ src/pathfinder/pathfinder_type.h
@@ -11,7 +11,7 @@
 #define PATHFINDER_TYPE_H
 
 #include "../tile_type.h"
-#include "npf/aystar.h"
+#include "aystar.h"
 
 /** Length (penalty) of one tile with YAPF */
 static const int YAPF_TILE_LENGTH = 100;
