+++ src/pathfinder/queue.cpp
